+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog for CardDAV client library for PHP ("PHP-CardDavClient")
 
+## Since last version
+- Support for OAUTH2/Bearer authentication. Specify bearertoken in credentials when creating Account. Acquiring the
+  access token is outside the scope of this library.
+- The interface for specifying credentials for an Account changed. The old username/password parameters are deprecated,
+  but still work.
+
 ## Version 1.1.0 (to 1.0.0)
 
 - New API AddressbookCollection::query() for server-side addressbook search